client/requirements.txt
grpcio==1.62.1  
grpcio-tools==1.62.1  
kafka-python==2.0.2  